Use 64-bit floats in `float_arrays`
Some tests where failing due to of the mix use of 64- and 32- bit
floats.

+++ invisible_cities/core/testing_utils.py
@@ -73,7 +73,7 @@ def float_arrays(size       =   100,
                       **kwargs)
     if mask is not None:
         elements = elements.filter(mask)
-    return arrays(dtype    = np.float32,
+    return arrays(dtype    = np.float64,
                   shape    =       size,
                   elements =   elements,
                   unique   =     unique)
